feat(postgres): migrate from in memory storage to postgres

+++ server/app/src/main/kotlin/mu/muse/persistence/instrument/inmemory/InMemoryInstrumentRepository.kt
@@ -36,7 +36,7 @@ class InMemoryInstrumentRepository(
         storage[instrument.id] = instrument
     }
 
-    override fun totalElements(): Int = storage.values.size
+    override fun totalElements(): Long = storage.values.size.toLong()
 }
 
 @Suppress("CyclomaticComplexMethod")
